added menu and moved indexes
