+++ apps/samradsgatt/components/About/index.ts
@@ -1,0 +1,3 @@
+export { default as Information } from './Information'
+export { default as MainColumn } from './MainColumn'
+export { default as RightSideColumn } from './RightSideColumn'
