+++ apps/samradsgatt/components/CaseDetails/LeftSideColumn/LeftSideColumn.tsx
@@ -7,7 +7,7 @@ import {
   Breadcrumbs,
 } from '@island.is/island-ui/core'
 import CaseTimeline from './CaseTimeline'
-import SubscriptionBox from '../../SubscriptionBox'
+import SubscriptionBox from '../../SubscriptionBox/SubscriptionBox'
 import { useLocation } from 'react-use'
 
 interface LeftSideColumnProps {
