+++ apps/samradsgatt/components/index.ts
@@ -1,0 +1,2 @@
+export * from './About'
+export * from './Card/Card'
